Move `pants.backend.java.tailor` -> `pants.backend.java.goals.tailor`.
[ci skip-rust]

[ci skip-build-wheels]
