+++ src/python/pants/backend/java/BUILD
@@ -2,5 +2,3 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_sources()
-
-python_tests(name="tests")
